scripts/test-e2e-parallel: split the runner into focused modules
Pure code-move, no behavior change. The ~1700-line runner is split into
scripts/e2e-parallel/{constants,argParser,scheduler,taskDiscovery,
nodeInfra,logging,runTask}.js, leaving test-e2e-parallel.js as the main()
orchestration entry point. The scheduler — the thread-budget cost model
plus the load-average-per-core adaptive controller — is now its own file,
reviewable in isolation.

Verified --dry-run output is byte-identical across default / --shared-node
/ --per-slot-node, tsc clean, and functional runs (default in-process and
--shared-node) pass unchanged.
